last commit

+++ src/main/java/com/example/sb1/entity/Person.java
@@ -2,6 +2,8 @@ package com.example.sb1.entity;
  
 import org.springframework.data.annotation.Id;
  
-public record Person(@Id Long id, String name, String email) {
+public record Person(@Id int id, String name, String email) {
+
+  
 }
